refactor: make shisen_cpp dir in src and change example to viewer
