+++ src/shisen_cpp/camera/node/camera_node.cpp
