+++ src/shisen_cpp/camera/provider/camera_config_provider.cpp
