+++ src/shisen_cpp/camera/provider/image_provider.cpp
