+++ src/shisen_cpp/utility/capture_setting.cpp
