+++ src/shisen_cpp/utility/mat_image.cpp
